ctr: move events command
